% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_functions.R
\name{get_progression_times}
\alias{get_progression_times}
\title{Sample date-time values from fire progression polygons}
\usage{
get_progression_times(
  prog,
  datetimecol,
  sampline,
  sampline_idcols = c("refid", "line_index", "segment"),
  return_geom = TRUE
)
}
\arguments{
\item{prog}{An \code{sf} spatial data frame of fire progression polygons. The
data must have a coordinate reference system defined and the progression
polygons must not overlap (both of these conditions will be checked).}

\item{datetimecol}{(character) The name of the column in the input
\code{progressions} data frame with date-time values. The date-times must
be \code{POSIX} values (e.g. class \code{POSIX.ct}).}

\item{sampline}{An \code{sf} spatial data frame of sampling lines as
generated by function \code{make_sample_lines}.}

\item{sampline_idcols}{(character) The name of one or more columns that
identify each sample line and that should be copied from the sample lines
layer to the output data frame. Defaults to \code{c("refid", "line_index",
"segment")} which are the column names used by function \code{make_sample_lines()}.}
}
\value{
Either an \code{sf} spatial data frame with \code{LINESTRING}
  geometry if argument \code{return_geom} is \code{TRUE}, or a plain data
  frame if \code{return_geom} is \code{FALSE}. Each record corresponds to a
  line segment (part or whole of a sample line) that either intersects a
  progression polygon or is outside any polygons. Other columns are the
  sample line ID attributes, progression date-time, segment distance
  (distance from start of sample line), and segment length.
}
\description{
Sample date-time values from fire progression polygons
}
\examples{
\dontrun{
library(CERMButils)
library(sf)

# Load a layer of back-burning lines and generate sample lines with the
# desired maximum length and spacing
dat_bb <- st_read("path/to/bb_lines")
dat_slines <- make_sample_lines(dat_bb, bb_id = "OBJECTID", trim_by = "all")

# Load a corresponding layer of fire progression polygons and get
# the date-time value(s), and possibly gaps of no data, along each sample line
dat_prog <- st_read("path/to/progressions")

dat_prog_samples <- get_progression_times(dat_prog,
                                          datetimecol = "LASTTIME",
                                          dat_slines)
}

}
\seealso{
\code{\link{make_sample_lines}}
}
