% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_progressions.R
\name{make_progressions}
\alias{make_progressions}
\title{Derive a set of progression polygons}
\usage{
make_progressions(x, time_cols, min_geom_area = 100)
}
\arguments{
\item{x}{An \code{sf} spatial data frame with one or more columns defining time
and polygonal features representing the total extent of a fire at each time.}

\item{time_cols}{A character vector of one or more column names that define
the time steps for the progressions.}

\item{min_geom_area}{Threshold minimum area (square metres) for progression
polygons. If the progression for a given time step consists of two or more
spatially distinct parts, the threshold is applied to each individually.}
}
\value{
An \code{sf} spatial data frame of progression polygons with
  attribute values copied from the input data.
}
\description{
This function takes a spatial data frame of fire extent polygons with one or
more attributes defining time points, and derives the corresponding sequence
of fire progression polygons.
}
