% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{remove_duplicate_geoms}
\alias{remove_duplicate_geoms}
\title{Remove features with duplicate geometries}
\usage{
remove_duplicate_geoms(x, ...)
}
\arguments{
\item{x}{An \code{'sf'} spatial data frame.}

\item{...}{Optional further arguments to specify the ordering for each group
of features with identical geometries. Ordering is specified using the same
syntax as with the \code{dplyr} \code{\link[dplyr]{arrange}} function.}
}
\description{
Line data for back-burns often contains a large number of exact spatial
duplicates, i.e. the same geometry. This function identifies groups of such
duplicates and subsets each group to a single feature by discarding all but
the first feature in the group. Optionally, other line attributes can be used
to order the line features before selecting the feature to retain by
supplying further arguments using the same bare variable name form as used
with the \code{arrange} function in the \code{dplyr} package (see example
below).
}
\details{
This function can also be used for other types of features, e.g. polygons or
points.
}
\examples{
\dontrun{
  library(CERMButils)

  # Identify groups of features with identical geometries and retain a
  # single feature for each group
  dat_firelines_unique <- remove_duplicate_geoms(dat_firelines)

  # For each group of features, retain the one with the latest
  # date-time value
  dat_firelines_latest <- remove_duplicate_geoms(dat_firelines, desc(ADDEDDATETIME))

  # Alternatively, retain the feature with the earliest date-time value
  dat_firelines_earliest <- remove_duplicate_geoms(dat_firelines, ADDEDDATETIME)
}

}
