% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_progressions.R
\name{make_progressions}
\alias{make_progressions}
\title{Derive a set of progression polygons}
\usage{
make_progressions(
  x,
  time_cols,
  extra_cols = NULL,
  cookie_cutter = NULL,
  out_epsg = 8058,
  min_geom_area = 100,
  unique_geom = TRUE,
  replicate_times = c("merge", "largest", "smallest", "fail"),
  dTolerance = 2
)
}
\arguments{
\item{x}{An \code{sf} spatial data frame with one or more columns defining time
and polygonal features representing the total extent of a fire at each time.}

\item{time_cols}{A character vector of one or more column names that define
the time steps for the progressions. Normally a time column will will hold
a Date or Date-Time value (e.g. POSIX.ct) but any scalar variable type that
can be sorted is acceptable. When two or more time columns are specified,
sorting will be based on the provided order of column names.}

\item{extra_cols}{An optional character vector with the names of additional
columns to retain in the output layer. The default (\code{NULL}) means no
additional columns.}

\item{cookie_cutter}{An optional \code{sf} spatial data frame of polygonal
features that define the outer-most supposed extent of the fire. If
provided, progression polygons will be clipped and or removed so that none
extend outside the cookie features.}

\item{out_epsg}{Integer EPSG code specifying the map projection for the
output progression polygons. This \emph{must} be a projected coordinate
system with metres as map units. The default is 8058 (NSW Lambert /
GDA2020).}

\item{min_geom_area}{Threshold minimum area (square metres) for progression
polygons. If the progression for a given time step consists of two or more
spatially distinct parts, the threshold is applied to each individually.}

\item{unique_geom}{(logical) If \code{TRUE} (the default), any group of
records with identical geometries will be reduced to a single record, being
that with the earliest time in the group. If \code{FALSE} this step will be
skipped.}

\item{replicate_times}{A single string that specifies what to do if there is
more than one input data record for any time step. Options are:
\code{'merge'} (merge feature geometries within each time step);
\code{'largest'} (choose the record with the largest extent polygon area);
\code{'smallest'} (choose the record with the largest extent polygon area);
\code{'fail'} (stop with an error message).
The default is \code{'merge'}. The argument may be abbreviated and is
case-sensitive.}

\item{dTolerance}{Distance tolerance used to simplify the input extent
polygons using the \code{\link[sf]{st_simplify}} function. Default is 2 for
a minimum between vertex distance of 2 metres. Set to 0 to omit the
simplification step.}
}
\value{
An \code{sf} spatial data frame of progression polygons with
  time column(s) copied from the input fire extent data.
}
\description{
This function takes a spatial data frame of fire extent polygons with one or
more attributes defining time points, and derives the corresponding sequence
of fire progression polygons.
}
